Refactor - OOP Principle
Refactor Duke.java to better structure the Bot with OOP principle.

Strip Duke.java into 4 components:
1. Bot.java - imitate a whole bot's body, which currently contains the Bot's Brain
2. BotBrain.java - imitate a bot's brain, which handles the interactions with the user
3. BotPrinter.java - imitate a bot's information output to the user
4. BotMemory.java - imitate a bot's memory unit, where it stores the commonly used lines
