+++ src/main/java/Bot.java
@@ -1,0 +1,12 @@
+import java.time.Period;
+import java.util.Arrays;
+import java.util.Scanner;
+import java.util.stream.IntStream;
+
+public class Bot {
+
+    public static void main(String[] args) {
+        new BotBrain().initiate();
+    }
+
+}
